requirements/base.txt
aioredis==2.0.0
asgi-lifespan==1.0.1
dnspython==2.1.0
fastapi==0.70.0
fastapi-cache2==0.1.7
httpx==0.21.1
pytest-asyncio==0.16.0
python-dotenv==0.19.2
uvicorn==0.15.0
